#270 feat : Category Selector Component reuse
