+++ src/components/mypick/myplace/Myplace.jsx
@@ -9,7 +9,8 @@ import nothingIcon from "../../../assets/img/nothing.svg";
 import { useNavigate } from "react-router-dom";
 import Request from "../../../functions/common/Request";
 import ChangeMode from "../../../assets/img/Mypick/ChangeMode.svg"
-import { CATEGORY_LIST, MatchCategory } from "../../common/Category";
+import CategorySelector, { CATEGORY_LIST, MatchCategory } from "../../common/Category";
+
 const Container = styled.div`
   margin: 0 auto;
   margin-top: 3%;
@@ -62,47 +63,13 @@ const ChangeModeButton = styled.span`
     top: 0.5%;
     font-size: 1rem;
   }
+  z-index: 3;
 `
-const CategoryCheckBox = styled.div`
+const FilterOptions = styled.div`
   position: absolute;
-  right: 10vw;
-  top: -5%;
-  display: flex;
-`;
-const CategoryLabelWrapper = styled.div`
-  width: 100%;
-  box-sizing: border-box;
-  display: flex;
-  justify-content:center;
-  align-items: center;
-  padding: 1%;
-`;
-const CategoryLabel = styled.div`
-  display: flex;
-  flex-flow: column wrap;
-  align-items: center;
-  text-align: center;
-  font-size: 12px;
-`
-const CategoryImageWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: 1px #99a0b0 solid;
-  border-radius: 50%;
-  width: 4vw;
-  height: 4vw;
-  @media screen and (max-width: 768px) {
-    width:10vw;
-    height: 10vw;
-  }
-`
-const CategoryNameWrapper = styled.div`
-  margin-top: 5%;
-  font-size: 0.65rem;
-  @media screen and (max-width: 768px) {
-    font-size: 0.65rem;
-  }
+  right: 10%;
+  top: -50%;
+  width: 30%;
 `
 
 const Myplace = (props) => {
@@ -162,45 +129,12 @@ const Myplace = (props) => {
               <img src={ChangeMode} style={{ marginRight: '10px' }} />
               STORY
             </ChangeModeButton>
-            <span style={{ fontWeight: "500", fontSize: "1.6rem", color: "#000000" }}>
+            <span style={{width:"100%",textAlign:'center', fontWeight: "500", fontSize: "1.6rem", color: "#000000", position:'relative' }}>
               MY PLACE
+              <FilterOptions>
+                <CategorySelector checkedList={checkedList} onCheckedElement={onCheckedElement}/>
+              </FilterOptions>
             </span>
-            <CategoryCheckBox>
-              {CATEGORY_LIST.map((item) => {
-                return (
-                  <CategoryLabelWrapper key={item.id}>
-                    <input
-                      type="checkbox"
-                      // 이때 value값으로 data를 지정해준다.
-                      value={item.data}
-                      // onChange이벤트가 발생하면 check여부와 value(data)값을 전달하여 배열에 data를 넣어준다.
-                      onChange={(e) => {
-                        onCheckedElement(e.target.checked, e.target.value);
-                        if (e.target.checked) {
-                          e.target.closest('div').style.color = 'red';
-                        }
-                        else {
-                          e.target.closest('div').style.color = 'black';
-                        }
-
-                      }}
-                      // 체크표시 & 해제를 시키는 로직. 배열에 data가 있으면 true, 없으면 false
-                      checked={checkedList.includes(item.data) ? true : false}
-                      id={`category${item.id}`}
-                      style={{ display: 'none' }}
-                    />
-                    <label htmlFor={`category${item.id}`}>
-                      <CategoryLabel>
-                        <CategoryImageWrapper>
-                          <img src={require(`../../../assets/img/Category/Category${item.id}.svg`)} style={{ width: '60%' }} />
-                        </CategoryImageWrapper>
-                        <CategoryNameWrapper>{item.name}</CategoryNameWrapper>
-                      </CategoryLabel>
-                    </label>
-                  </CategoryLabelWrapper>
-                );
-              })}
-            </CategoryCheckBox>
             <main style={{ width: '100%'}}>
               <Container>
                   {info.length === 0 ? (
